Parallelism: Add basic solution for HW1
Added a solution that implements a parallel odd-even merge sort.
The idea behind this algorithm is leveraging multiple cores to quickly
perform a merge sort over a list of integer.
We use OpenMP to implement the parallelism.

Currently the code is relatively slow (much slower than just normal
merge sort), maybe because of cache misses or some misuse of the OpenMP
library.
Regardless, the code can be easily changed into a regular version of a
merge sort, albeit with a few usages of the OpenMP library (specifically
for operations which can be done in parallel).